requirements.txt
fastapi
uvicorn
jinja2
langgraph
langchain-core
opentelemetry-api
chromadb
python-dotenv
openai
requests
redis
psutil
confluent-kafka
